implement payroll services

+++ TeaPlanter/src/main/java/online/jayashan/teaplanter/services/payrollService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface payrollService {
     payroll createNewStudent (payroll payroll);
     List<payroll> getAllStudents();
-    payroll getStudentById(Integer id);
-    payroll updateStudentById(Integer id, payroll updatedStudent);
-    void deleteStudent(Integer id);
+    payroll getStudentById(Long id);
+    payroll updateStudentById(Long id, payroll updatedStudent);
+    void deleteStudent(Long id);
 }
